Issue #63
Adding shown and hidden to the Dockable and DockingListener interfaces. Shown is called when the dockable is shown (but not when it is docked, that is a separate call). Hidden is called when the dockable is hidden, but not undocked. A current exception is when a dockable is closed that is in a tabbed pane, hidden and undock will both be called in this case.