Issue #54
Moving check for dockable to the removePanel method in DockedTabbedPanel in case it is called from outside the class in the future.

Both base Docking.dock methods now undock the dockable if it is currently docked. This ensures that we don't steal the UI components from other panels. This was causing issues with the tab count and panel count not matching in DockedTabbedPanel.