#83
Removing check that forces x position of floating frame to be positive. This value might be negative if the screens are not in order.

I guess in theory this could also be true for monitors that are stacked on top of each other. A more general solution for making sure the dockable is on screen should be found.