Issue #103
Adding isWrappableInScrollpane to Dockable. By default, this is true and the docking framework will wrap the Dockable in a JScrollPane when adding it to the layout. Overriding isWrappableInScrollpane and return false will prevent this. False is recommended if the panel contains a JTable as that will also have a JScrollPane and cause issues.