Floating rework (#119)
Floating has been reworked to use JPanels (DockedSimplePanel and DockedTabbedPanel) instead of the dockables directly. This allows the user to float individual dockables or an entire tabbed panel of dockables. In theory this opens up the possibility of floating other types of panels. But there are no plans right now.