Issue #196
Dynamic dockables in tab groups now reload properly. This required storing the class-name for the selected tab and all tabs in the group pane. This allows Modern Docking to properly reload. There were also some code changes required to detect when a dockable has failed to load, and to check if it is a dynamic dockable (by attempting to create an instance of the class).