Issue #37 Issue #286
Modern Docking has a new listener option for detecting when new floating frames are created. This new floating frame listener allows the application to modify the new floating frames as they see fit.

Modern Docking is still the one creating the frame, so this is naturally limited a little. The application won't be able to create their own frame in its place. The default close operation should not be modified and left as `DISPOSE_ON_CLOSE`. Modern Docking will still dispose these frames when they are empty.