Custom Hook for Dynamic Dockable Creation (#360)
* Issue #355

API users are now able to add their own listener to create dynamic dockables when they are loaded from a layout.

Some documentation is needed and it would be nicer to have a better test setup for this and other dynamic dockable features.

* Issue #355

Add properties to the dynamic dockable callback.